[IMP] mail: move discuss app model to the correct bundle
The mail module is split into several sub-folders to determine when
the files should be loaded (e.g., web, public, public_web...).

Currently, several code snippets are misplaced, which increases the
load on public pages or live chat.

Moreover, this can lead to dependency hell, especially when moving a
feature to another bundle.

This PR moves the discuss app model, which is only used in the web
and public bundle, to the public_web bundle.

Part of task-3972988
