+++ addons/im_livechat/__manifest__.py
@@ -70,6 +70,7 @@ Help your customers with this chat, and analyse their feedback.
             'im_livechat/static/src/scss/im_livechat_history.scss',
             'im_livechat/static/src/scss/im_livechat_form.scss',
             'im_livechat/static/src/core/common/**/*',
+            'im_livechat/static/src/core/public_web/**/*',
             'im_livechat/static/src/core/web/**/*',
         ],
         'web.assets_unit_tests': [
