+++ addons/mail/static/src/core/public_web/discuss_app_model.js
@@ -1,5 +1,5 @@
+import { Record } from "@mail/core/common/record";
 import { _t } from "@web/core/l10n/translation";
-import { Record } from "./record";
 
 export class DiscussApp extends Record {
     static new(data) {
