+++ addons/mail/static/src/core/public_web/store_service_patch.js
@@ -1,0 +1,34 @@
+import { Store, storeService } from "@mail/core/common/store_service";
+import { Record } from "@mail/core/common/record";
+import { router } from "@web/core/browser/router";
+import { patch } from "@web/core/utils/patch";
+
+patch(Store.prototype, {
+    setup() {
+        super.setup(...arguments);
+        this.discuss = Record.one("DiscussApp");
+        this.action_discuss_id;
+    },
+    onStarted() {
+        super.onStarted(...arguments);
+        this.discuss = { activeTab: "main" };
+    },
+});
+
+patch(storeService, {
+    start(env, services) {
+        const store = super.start(...arguments);
+        const discussActionIds = ["mail.action_discuss", "discuss"];
+        if (store.action_discuss_id) {
+            discussActionIds.push(store.action_discuss_id);
+        }
+        store.discuss.isActive ||= discussActionIds.includes(router.current.action);
+        services.ui.bus.addEventListener("resize", () => {
+            store.discuss.activeTab = "main";
+            if (services.ui.isSmall && store.discuss.thread?.channel_type) {
+                store.discuss.activeTab = store.discuss.thread.channel_type;
+            }
+        });
+        return store;
+    },
+});
