+++ addons/mail/static/src/core/public_web/thread_model_patch.js
@@ -1,0 +1,58 @@
+import { patch } from "@web/core/utils/patch";
+import { Thread } from "@mail/core/common/thread_model";
+import { Record } from "@mail/core/common/record";
+import { router } from "@web/core/browser/router";
+
+patch(Thread.prototype, {
+    setup() {
+        super.setup(...arguments);
+        this.discussAppCategory = Record.one("DiscussAppCategory", {
+            compute() {
+                return this._computeDiscussAppCategory();
+            },
+        });
+    },
+
+    _computeDiscussAppCategory() {
+        if (["group", "chat"].includes(this.channel_type)) {
+            return this.store.discuss.chats;
+        }
+        if (this.channel_type === "channel") {
+            return this.store.discuss.channels;
+        }
+    },
+
+    /** @param {boolean} pushState */
+    setAsDiscussThread(pushState) {
+        if (pushState === undefined) {
+            pushState = this.notEq(this.store.discuss.thread);
+        }
+        this.store.discuss.thread = this;
+        const activeId =
+            typeof this.id === "string" ? `mail.box_${this.id}` : `discuss.channel_${this.id}`;
+        this.store.discuss.activeTab =
+            !this.store.env.services.ui.isSmall || this.model === "mail.box"
+                ? "main"
+                : ["chat", "group"].includes(this.channel_type)
+                ? "chat"
+                : "channel";
+        if (pushState) {
+            router.pushState({ active_id: activeId });
+        }
+    },
+
+    async unpin() {
+        this.isLocallyPinned = false;
+        if (this.eq(this.store.discuss.thread)) {
+            router.replaceState({ active_id: undefined });
+        }
+        if (this.model === "discuss.channel" && this.is_pinned) {
+            return this.store.env.services.orm.silent.call(
+                "discuss.channel",
+                "channel_pin",
+                [this.id],
+                { pinned: false }
+            );
+        }
+    },
+});
