+++ addons/mail/static/src/core/web/@types/models.d.ts
@@ -11,6 +11,7 @@ declare module "models" {
         activityCounter: number,
         activity_counter_bus_id: number,
         activityGroups: Object[],
+        getDiscussSidebarCategoryCounter: (categoryId: number) => number,
     }
     export interface Thread {
         recipients: Follower[],
