+++ addons/mail/static/src/core/web/chat_hub_patch.js
@@ -1,0 +1,8 @@
+import { patch } from "@web/core/utils/patch";
+import { ChatHub } from "@mail/core/common/chat_hub";
+
+patch(ChatHub.prototype, {
+    get isShown() {
+        return super.isShown && !this.store.discuss.isActive;
+    },
+});
