+++ addons/mail/static/src/core/web/store_service_patch.js
@@ -142,6 +142,15 @@ const StorePatch = {
             systray_get_activities: true,
         };
     },
+    getDiscussSidebarCategoryCounter(categoryId) {
+        return this.DiscussAppCategory.get({ id: categoryId }).threads.reduce((acc, channel) => {
+            if (categoryId === "channels") {
+                return channel.message_needaction_counter > 0 ? acc + 1 : acc;
+            } else {
+                return channel.selfMember?.message_unread_counter > 0 ? acc + 1 : acc;
+            }
+        }, 0);
+    },
     getNeedactionChannels() {
         return this.getRecentChannels().filter((channel) => channel.importantCounter > 0);
     },
